+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * fix(commands): display help usage message if needed [#835](https://github.com/Scalingo/cli/pull/835)
 * feat(logs): allow DB type aliases [#829](https://github.com/Scalingo/cli/pull/829)
 * feat(db-console): add `influx-console` and `mongodb-console` aliases [#830](https://github.com/Scalingo/cli/pull/830)
-
+* feat(send-signal): `send-signal` command added [#833](https://github.com/Scalingo/cli/pull/833)
 ### 1.26.1
 
 * deps(urfave/cli): Bumpd from 2.17.1 to 2.23.5: fix options display regression in --help [#814](https://github.com/Scalingo/cli/pull/814)